Move ConvertToExtensionMethod tool
